refactor: invert dependency between features and infra:core
