+++ features/terms/list/build.gradle
@@ -49,7 +49,6 @@ dependencies {
 
     implementation project(":features:core:di")
 
-    implementation project(":infra:core:repository")
     implementation project(":infra:contracts")
 
     implementation project(":foundation:provider")
