+++ infra/contracts/src/main/java/com/example/contracts/usecase/IGetUserUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.contracts
+package com.example.contracts.usecase
 
 import com.example.models.User
 import kotlinx.coroutines.flow.Flow
