+++ infra/core/repository/build.gradle
@@ -31,6 +31,9 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4"
+
+    implementation project(":infra:contracts")
+
     implementation project(":foundation:models")
     implementation project(":foundation:cache")
 }
